+++ cmd/action/create/cluster/defaultcontrolplane/runner.go
@@ -1,4 +1,4 @@
-package onenodepool
+package defaultcontrolplane
 
 import (
 	"context"
